Touch gestures list for using features in text editor
The file contains detailed touch gesture list and instruction to perform features in text editor. Each feature has a distinct gesture pattern for alphabetic and symbolic gesture. For example to select all texts, draw A on screen for text editor with alphabetic gesture. To select all text in application with symbolic gesture, draw a rectangle.